Refactor TodoItem code to class component
