+++ src/components/TodoItem.js
@@ -1,43 +1,71 @@
+import React, {Component} from 'react';
+
 // css styles module
 import styles from "./TodoItem.module.css"
 
-function TodoItem(props) {
-  const completedStyle = {
-    fontStyle: "italic",
-    color: "#595959",
-    opacity: 0.4,
-    textDecoration: "line-through",
-  }
+class TodoItem extends Component {
+    state = {
+      editing: false,
+    }
 
-  const {completed, id, title } = props.todo;
+    handleEditing = () => {
+      this.setState({
+        editing: true,
+      })
+    }
+    
+    handleUpdatedDone = (event) => {
+      if (event.key === "Enter") {
+        this.setState({ editing: false })
+      }
+    }
+
+  render(){
+    let viewMode = {};
+    let editMode = {};
+
+    if (this.state.editing) {
+      viewMode.display = "none"
+    } else {
+      editMode.display = "none"
+    }
+    const completedStyle = {
+      fontStyle: "italic",
+      color: "#595959",
+      opacity: 0.4,
+      textDecoration: "line-through",
+    }
+
+    const {completed, id, title } = this.props.todo;
   
-  return (
-    <li className={styles.item}>
-      <div onDoubleClick={props.handleEditing} style={props.viewMode}>
-        <input
-        type="checkbox"
-        className={styles.checkbox}
-        checked={completed}
-        onChange={() => props.handleChangeProps(id)} />
-        <button onClick={() => props.deleteTodoProps(id)} >
-          Delete
-        </button>
-        <span style={completed ? completedStyle : null}>
-          {title}
-        </span>
-      </div>
-      <input 
-        type="text" 
-        style={props.editMode}
-        value={title}
-        className={styles.textInput} 
-        onChange={e => {
-          props.setUpdate(e.target.value, id)
-        }}
-        onKeyDown={props.handleUpdatedDone}
-      />
-    </li>
-  )
+    return (
+      <li className={styles.item}>
+        <div onDoubleClick={this.handleEditing} style={viewMode}>
+          <input
+          type="checkbox"
+          className={styles.checkbox}
+          checked={completed}
+          onChange={() => this.props.handleChangeProps(id)} />
+          <button onClick={() => this.props.deleteTodoProps(id)} >
+            Delete
+          </button>
+          <span style={completed ? completedStyle : null}>
+            {title}
+          </span>
+        </div>
+        <input 
+          type="text" 
+          style={editMode}
+          value={title}
+          className={styles.textInput} 
+          onChange={e => {
+            this.props.setUpdate(e.target.value, id)
+          }}
+          onKeyDown={this.handleUpdatedDone}
+        />
+      </li>
+    )
+  }
 }
 
 export default TodoItem;
